refactor: select passport tests
